requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.2
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.0
h11==0.14.0
hiredis==2.2.3
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
numpy==1.26.1
orjson==3.9.7
pandas==2.1.1
protobuf==4.21.12
psycopg==3.1.12
psycopg-binary==3.1.12
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==3.0.0
